GEODE-3052 Restarting locators causes potential locator split brain
When restarting from a locatorView.dat file we should ignore any locator
entries in the view.  Recovery tries to get this state from other locators
before resorting to using the persisted view so there we know all of the
locator entries in the view are invalid.  This allows the locators to
quickly move into the concurrent-startup algorithm and find each other.

I removed the Flaky categorization of the test that I modified to
reproduce the problem.  A subclass's use of the test was reported as
a Flaky failure but I found that the ticket was closed.