GEODE-3172: Fix serialization errors copying queue between 1.0 and 1.2
Deserialize a HAEventWrapper using the version of the sender when
receiving a GII.

Serialize entries using the version of the remote member when sending
data as part of GII. This works for the client queues because client
queues always have deserialized values. If there is an internal region
that has serialized values in memory, those values would still be copied
on the wire directly without being translated to the old members
version.

Adding a test that demonstrates the serialization issues we were seeing
with this issue. The test starts a 1.0 server, puts some data in the
queue and starts a 1.2 server.

This closes #620