GEODE-1731: Modifying region size check to be more accurate
Changed the hardcoded 5010 region size check to the size of the region already plus 20 new elements. The 20 new elements is due to the fact that a synchronous call to get the region size of vm2 had to be used, so there should be more time for elements to be put.