GEODE-3333: GMSJoinLeaveJUnitTest.testViewNotSentWhenShuttingDown fails
Moved assertion to remove race condition.  The test wanted to
confirm that the object under test thought it was membership
coordinator and then check to see if it decided to stop being
coordinator.  If the background thread did this quickly enough
the first condition would never be seen by the test.

Since the object under test was the coordinator immediately after
setup there was no need to delay the assertion until after removing
other members from the view.