GEODE-3306: Fix unit test
Remove previous setup code for `parse()`, which is already called in
`CacheFactory.create()`.

The end result is a simpler `CacheFactory.create()` call. The path of
`cache.xml` is different because the `GemFireCacheImpl` class is the one
that retrieves `cache.xml`, but its package is
`org.apache.geode.internal.cache` and not
`org.apache.geode.internal.cache.xmlcache` (where this class resides).

Hence the additional `xmlcache` is needed for `GemFireCacheImpl` to find
the resource for the unit test.