GEODE-3047 Atomic creation flag is not set if the region is recoverd from the disk.
While creating bucket region, to satisfy the redudndancy copies the bucket regions
are created on all vailable nodes. The initialization (setting persistentIDs) of
these buckets are done after creating buckets on all the nodes. This introduced
a race condition for the bucket region which are recovered from the disk to
exchange thier old id with the peer node resulting in ConflictingPersistentData
Exception.

The changes are done so that the regions persistent ids are set as soon as they
are created/initialized.