GEODE-2708: Update Gradle Wrapper To 3.5.1
While updating the Gradle wrapper from 2.14.1 to 3.5.1, it was found that the
ANTLR plugin causing an issue inside the Eclipse IDE. The fix was to declare the
ANTLR dependency version for the antlr configuration. If no version is set for
the antlr plugin, it will default to 2.7.7. However, this was causing an issue
for Eclipse and other IDE's because the plugin was modifying the antlr
configuration after the initial dependency resolution was done.

- Updated Gradle Wrapper to 3.5.1
- Added antlr dependency to the antlr configuration

Verified by importing project into Eclipse and also running build from
command-line