GEODE-3529 move new client/server security classes to a different package
The new classes/interfaces have been moved to org.apache.geode.security.server.

A package.html is now provided that describes the purpose of the package.

The word "Stream" has been removed from the names of these classes.  Those
names make sense in the context of the Protobuf support classes in
org.apache.geode.internal.cache.tier.sockets classes but not when taken
as a separate group of classes.

The method Authenticator.receiveMessage() has been renamed to
Authenticator.authenticate().

This closes #746