GEODE-3059: LoadMonitor.connectionClosed incrementing statistics only for client-server connection
The stat should only be incremented/decremented for certain kinds of
connections.  I've modified it to include protobuf connections.  All
of the constant byte identifiers in Acceptor.java have been moved to
an enum in CommunicationMode.java.  In that class I've added some
"isa" checks to replace the many big "if" checks for different kinds
of connection modes.

A new connection modes will henceforth need to be added to
CommunicationMode.java where the appropriate "isa" methods can be
updated to include the new mode.