GEODE-3580: patch test to avoid the current failure.
When a cache has already been created and closed in the current JVM,
`TypeRegistry` will have a flag set that stops a distributed system from
sending messages. This means we can't connect to a distributed system,
even though there doesn't seem to be any reason to disallow it. This is
to stop the test failing; a real fix will come later.