Adding configuration options for TestRunners and removing warmup
Adding options to configure which concurrent test runner to use, and
adding configuration for the JPF runner to pass jpf properties.

Adding an alternative runner which just runs the test in a loop.

Removing the warmup from the FilterProfileTest. It turns out most of the
work was in calling static initializers, especially for logging classes.
Disabling choice generation in static final initialization, and
disabling the logger entirely.