Adding a new framework for concurrency testing in geode (#787)
* Adding a new framework for concurrency testing in geode

Added a new junit runner, ConcurrentTestRunner, for running a test that
has parallel threads.

The runner currently runs the test using Java PathFinder, which will
run the test with all interleavings of the threads.

Added an example test to geode-core for the FilterProfile serialization
logic.

* Adding configuration options for TestRunners and removing warmup

Adding options to configure which concurrent test runner to use, and
adding configuration for the JPF runner to pass jpf properties.

Adding an alternative runner which just runs the test in a loop.

Removing the warmup from the FilterProfileTest. It turns out most of the
work was in calling static initializers, especially for logging classes.
Disabling choice generation in static final initialization, and
disabling the logger entirely.