GEODE-3780 suspected member is never watched again after passing final check
A member going through a final health check is now put in the
suspected members collection and a new "neighbour" is selected for
the background monitor thread, ensuring that it doesn't interfere
with the health check.  Once the health check is done the member is
removed from the suspected members collection and a new "neighbour"
is selected, allowing the monitor thread to once again consider the
suspected member.