GEODE-3780 suspected member is never watched again after passing final check
incorporating review feedback from Hitesh.  We want members other than
the coordinator to also reconsider the suspected member.  The
Monitor will now invoke setNextNeighbor at the end of its run() method
and a final check that passes will result in a message being sent to
the initiator so that it can start watching the suspect again.