GEODE-3867 mutual SSL authentication with a bad client cert is not being tested
new test added & old test revised.  The old test had an odd try/catch
for an exception that should not be thrown.  The new test uses pretty
much that same structure & expects the client to fail to connect.  It
also expects that the server will not have received any updates from the
non-trusted client.