GEODE-3338 Refactor WAN command tests to use new Rules
This commit includes refactoring the 13 test classes in WAN Command tests to
use the new Test Rules vs. the deprecated CliCommandTestBase.
In addition to GEODE-3338, please refer to GEODE-3362 through GEODE-3374
(one per test plus one to remove WANCommandTestBase) all of which are included
as sub jiras for GEODE-1359.

- Remove use of deprecated APIs in CliCommandTestBase
- Remove WANCommandTestBase (all classes extended this class)
- Refactored to use new rules
- Removed FlakyTest specification for GEODE-1355, GEODE-1395, GEODE-1418
and GEODE-1448.