GEODE-3683 CacheConnectionTimeoutJUnitTest failing intermittently in CI
Increased the maximumTimeBetweenPings from 100ms to 2000ms.  Tests on
a mac with Intellij showed a minimum of 19ms between finishing the setup
method and commencing the actual test method.  Adding a short sleep at
the beginning of the test method caused it to fail repeatedly with the
lower setting.  A 2 second timeout period allowed the test to succeed
in 200 consecutive runs.
'