GEODE-3976: improve exception message
The message in the exception thrown when the disk monitor
detects we have gone critical (i.e. run out of space) will
now also tell why it went critical.
It will either say something like this:
the file system is 2 percent full, which exceeds the critical threshold of 1.9 percent.
or something like this:
the file system only has 98 bytes free which is below the minimum of 1048576.

Also added unit test for DiskUsage class.