GEODE-4101: Add --redirect-output flag to GFSH.
To avoid major code changes the start commands check the new flag and
set as `true` a new system property, `DISABLE_REDIRECTION_CONFIGURATION`,
which disables the other two (now deprecated) properties. In the next
major release, the properties and the related logic should be deleted.
The integration between log4j2 and geode-pulse was also incomplete.
When ran in embedded mode, some messages were totally swallowed and the
actual logging was split between the member's log and pulse.log file.

- Added flag `--redirect-ouput` to `StartServerCommand`.
- Added flag `--redirect-ouput` to `StartLocatorCommand`.
- Added tests to verify the behavior of the commands/builders.
- Added `OSProcess.DISABLE_REDIRECTION_CONFIGURATION` internal property.
- Property `OSProcess.ENABLE_OUTPUT_REDIRECTION` marked as deprecated.
- Property `OSProcess.DISABLE_OUTPUT_REDIRECTION` marked as deprecated.
- Added `log4j-jcl` as a dependency of `geode-pulse`. By adding this
  dependency no logging is lost and nothing is split: everything goes
  to the pulse.log file (or whatever is configured in `log4j2.xml`).