GEODE-4163 clean up DistributionManager code
Removed all dead code - about 800 lines of code have been removed.
Removed DistributionChannel, which served no useful purpose.
Made many methods private or package-private.
Gathered most statics and instance variables scattered around in the file
and moved them to the top.

I attempted to replace use of InternalDistributedMember with
DistributedMember but the use of the former is too widespread.  If we
want to do that it will change a large set of files.