GEODE-4178 ClientServerMiscDUnitTest.testOnSeverMethodsWithCachingProxyClientHA failure
This goofy test was setting a 1 second read-timeout and had only one server.
Our test JVMs have frequent GC cycles that can easily last over one second.

I could easily reproduce the failure by adding a cache writer on the server
that paused for a couple of seconds.

I've increased the read-timeout to 5 seconds.  It could be set to a longer
interval if this isn't enough.