GEODE-3861 ClientHealthMonitorSelectorJUnitTest.testDeadClientRemovalByServer failed
When Alexander tightened up the timing in this test in
b31de81a8c19c61ad27a2ab803073214050db154 he went a little too far.
The maximum time between pings was reduced from 2500ms to 50ms,
allowing the health monitor to timeout the connection for the client
in testDeadClientRemovalByServer before the test got to the point of using it.  Reversing the timeout change and increasing the wait time for the health monitor to timeout the client connection fixes the problem.

Reverting the change and increasing the wait time for the server to timeout
the thread fixes the problem.

This test wasn't marked flaky until the above changes were made so I'm
removing the flaky-test category from it and its superclass.