GEODE-4192 GetServer request should return error if no servers found
I modified the GetServerOperationHandler to return a failure if there are
no suitable servers.  There is a new NO_AVAILABLE_SERVER error code for
this.

I altered the test for this condition to require an error response.

I altered the experimental driver to look for the new error code and
added tests in DriverConnectionTest to ensure that it's working.