GEODE-4229 CI failure due to suspect string: "Locator socket was closed unexpectedly"
Removing error-level log message that was added during this cleanup
for GEODE-4176.

The server socket is closed as a matter of course in processing a
shutdown request, so we shouldn't consider this condition to be an
error.

(cherry picked from commit bb9ae49e0516d466d3e2665e952068afee0d502a)