GEODE-4148 Locator threshold before rebalancing clients
The server location service is now less aggressive in moving clients
from the most loaded server to the least loaded server.  It will wait
until the load difference is 10 times the load-per-connection value,
which defaults to 1.  Once this threshold is reached clients will be
moved until balance has been restored.

The threshold can be changed from 10 to some other floating point value
with the system property

   gemfire.locator-load-imbalance-threshold

when starting locators.  The setting is not needed in clients or
servers.