GEODE-3588 2 restarts of Locator results in split brain
Udo's fix for GEODE-870 added a new boolean instance variable to
GMSJoinLeave to tell its ViewCreator thread to shut down.  This works
but the state was never being reset after its first use.  This caused
Subsequent ViewCreator threads to shut down immediately.  The only
way to fix this condition without a patch is to restart the coordinator node.

The patch moves this boolean variable to the ViewCreator thread so that
it is automatically reset when a new ViewCreator is instantiated.

I also did a little code cleanup, moving GMSJoinLeave methods from the
end of the file to where its other methods are located and adding
a setShutdownFlag() method during debugging so I could isolate what
was happening.