GEODE-4101: Only check for specific command line arguments
StartServerCommandTest failed in CI build, but the assertion that failed
didn't provide any useful debugging information

Change the assertion to use assertj so more information will be lgged
regarding mismatches between expected and actual command arguments.

Also changed expectedCommandLineElements to only contain the command args that
we really care about - those that are specific to the command. Other arguments
maythat are dependant on the test environment may vary and ignored in
these tests.