GEODE-2556 RollingUpgradeDUnitTest.testRollServersOnPersistentRegion_dataserializable[0] fails intermittently
Most of the tests in this class use region shortcuts to create a Region
but those that use persistent regions were not & were failing to set
the Scope to DISTRIBUTED_ACK.  To compensate for the use of the default
DISTRIBUTED_NO_ACK scope the test had a threadSleep() method that paused
the test for 1/4 second hoping that distribution would finish and verification
would pass.

I've set the scope to DISTRIBUTED_ACK and have removed the threadSleep()
method.