GEODE-4317 add error handling to the experimental ProtobufDriver's Region implementation
I've added checks to ProtobufRegion for "error" responses from the server.
Along the way I found we were converting some keys to their string
representations and fixed that as well.

The exceptions thrown show the error text returned by the server.  The
handling of putAll responses was already looking for errors but wasn't
including the server's message text so I changed it to do so.

This closes #1322