GEODE-3948 Improve CQ performance under flaky network conditions
This adds a new PoolFactory setting allowing subscription connections to
time out and initiate failover to a backup server.  The new setting is
setSubscriptionTimeoutMultiplier:

 A server has an inactivity monitor that ensures a message is sent to a client at least once a
 minute (60,000 milliseconds). If a subscription timeout multipler is set in the client it
 enables timing out of the subscription feed with failover to another server.

 A value of zero (the default) disables timeouts

 A value of one will time out the server connection after one of its ping intervals (not
 recommended)

 A value of two or more will time out the server connection after that many ping intervals have
 elapsed

The client/server handshake is modified for clients having version 1.5 or
later.  The server sends its ping-interval setting to the client.  The client
then uses this and the multiplier to establish a read-timeout in the
CacheClientUpdater subscription processor.

Two tests are added to ensure that 1) the Message method that allows a
read to timeout functions correctly and 2) the CacheClientUpdater
correctly receives the multiplier setting.