GEODE-4487: Remove singleton calls from offheap package (#1382)
One test was using the singletons just to confirm the cache had been shutdown. It was easy to change that test to not confirm the singleton was null but to instead ask if the test's cache was closed.

The other test was testing DataType which is product code that is only used by tests to inspect what is stored in offheap memory. The test was mocking the singleton methods because the product DataType was using the singleton to decode a serialized PDX. I changed DataType to just do less decoding of pdx. Now it will just say it was a pdx and give you the type number.