GEODE-4615 Deadlock shutting down client cache
This modifies the synchronization in addConnection to perform a
shutdown check in ConnectionMap prior to allowing a new connection
to be added.  This prevents the lock-inversion seen in the deadlock.

I attempted to refactor ConnectionMap into a static class for easier
unit testing but it proved to be too intertwined with
ConnectionManagerImpl, so the new unit test had to be a little more
complicated than I would have liked.  I've opened GEODE-4617
to address refactoring this class.