GEODE-4615 Deadlock shutting down client cache
The previous fix had a bad flaw in that the thread closing the pool could
set closing=true after another thread has checked if closing==true and
obtained the connection map lock.

This revision replaces the connection map's connection list with a
poisoned list under sync and then closes each connection outside
of the connection map's lock.  This removes the lock inversion since
the closing thread no longer holds the map's lock.  The thread adding
a connection will either see closing==true or it will try to add
the connection to the poisoned list.  Either results in it
throwing a CacheClosedException.