GEODE-4377: Don't catch encoding exceptions in OperationHandlers
Addressing review comments.

I've added unit tests to ensure that a Success response is returned to
the client if there are any encoding/decoding errors encountered while
processing entries.

In ProtobufResponseUtilities I've modified the error message to contain
the toString() of the exception, which will include the exception class
name and the exception's message.