GEODE-4377: Don't catch encoding exceptions in OperationHandlers
Galen pointed out that the changes for this ticket changed behavior when
we get an encoding error.  The changes cause the connection to be closed
but we used to leave it open.

This commit reverts to the old behavior.  The connection will remain open
and the client can continue to do goofy things and get error responses.