GEODE-3523: Update locatorDiscoveryCallback after updating state
Some tests wait for this callback to be notified of new locators before
killing old locators. But The callback is called before the internal
state on the client is updated. So there was a race condition where the
test would know about the new locator, but the client code itself would
not use it yet.