GEODE-4672 Geode fails to start with JDK 9 if validate-serializable-objects is set
The ObjectInputFilter wrapper now uses reflection and a dynamic
proxy to interact with the serialization filter classes.  A dynamic
proxy is needed since we have to implement a method specified in
a class whose package has changed in the JDK.  The logic of the filter
has not been changed.

I also replaced a couple of recently added white-list wildcards with specific
classes and added a javadoc for the whitelist in InternalDataSerializer.
I opened a new JIRA ticket concerning another wildcard that was recently
added for the JDBC connector and left a TODO in place for this.